requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
unstructured[local-inference]==0.10.30

# LLM and AI frameworks
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.13
tiktoken==0.5.1

# Vector databases and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# Optional: Pinecone (cloud vector DB)
pinecone-client==2.2.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pandas==2.0.3
scikit-learn==1.3.0

# Frontend (optional)
streamlit==1.28.1

# Development and testing
pytest==7.4.3
httpx==0.25.2

# Logging and monitoring
loguru==0.7.2

aiohttp==3.9.5
